Update App.js
Added my part of the work, not really finished as there were problems with it. But for now when you inspect the console, you will at least see that it has managed to get data from the database but the problem is that it's providing the response data as "undefined".